add bidirectional support
